[template] MVC template - the controller and router
the initial template for implementing the MVC pattern.
separated the controller, which returns a model+view(=json), from the main express app, with a router.
the @Service and @Repository layer template will be worked on soon.

added dependencies 'prettier' for auto-formatting code, and the jest test framework with a sample api unit test.